fixed routing issues with auth pages
