+++ src/app/auth/forgotuser/page.tsx
@@ -1,18 +1,19 @@
+'use client';
+
 import { useState, useEffect } from "react";
+import { useRouter } from "next/router";
 import "./authStyles.css";
 
 interface Fields {
-	username: string;
 	email: string;
 	password: string;
 }
 
-function SignUp() {
-	const initialValues: Fields = { username: "", email: "", password: "" };
+function ForgotUser() {
+	const initialValues = { email: "", password: "" };
 
 	const [formValues, setFormValues] = useState<Fields>(initialValues);
 	const [formErrors, setFormErrors] = useState<Fields>(initialValues);
-
 	const [isSubmit, setIsSubmit] = useState(false);
 
 	const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -37,9 +38,6 @@ function SignUp() {
 		const errors: Fields = initialValues;
 		const regex = /^[^\s@]+@[^\s@]+\.[^\s@]{2,}$/i;
 
-		if (!values.username) {
-			errors.username = "Username is required!";
-		}
 		if (!values.email) {
 			errors.email = "Email is required!";
 		} else if (!regex.test(values.email)) {
@@ -52,33 +50,21 @@ function SignUp() {
 		} else if (values.password.length > 10) {
 			errors.password = "Password cannot exceed more than 10 characters";
 		}
-
 		return errors;
 	};
 
+	const router = useRouter();
+
+	function ResetUser() {
+		router.push("/resetuser");
+	}
+
 	return (
 		<div className="container">
-			{Object.keys(formErrors).length === 0 && isSubmit ? (
-				<div className="ui message success">Signed in successfully</div>
-			) : (
-				<pre>Login Unsuccesfull</pre>
-			)}
-
 			<form onSubmit={handleSubmit}>
-				<h1>Login Form</h1>
+				<h1>Reset Username Form</h1>
 				<div className="ui divider"></div>
 				<div className="ui form">
-					<div className="form">
-						<label>Username</label>
-						<input
-							type="text"
-							name="username"
-							placeholder="Username"
-							value={formValues.username}
-							onChange={handleChange}
-						/>
-					</div>
-					<p>{formErrors.username}</p>
 					<div className="form">
 						<label>Email</label>
 						<input
@@ -101,11 +87,11 @@ function SignUp() {
 						/>
 					</div>
 					<p>{formErrors.password}</p>
-					<button className="fluid ui button blue">Submit</button>
+					<button onClick={ResetUser}>Submit</button>
 				</div>
 			</form>
 		</div>
 	);
 }
 
-export default SignUp;
+export default ForgotUser;
