+++ src/app/header.tsx
@@ -1,11 +1,8 @@
 export default function Header() {
-    return (
-      <div>
-        <div>Investogram</div>
-          <body>
-          </body>
-      </div>
-    )
-  }
-
-  
+	return (
+		<div>
+			<div>Investogram</div>
+			<body></body>
+		</div>
+	);
+}
