+++ src/app/layout.tsx
@@ -1,23 +1,12 @@
-import './globals.css'
-import Header from './header'
-
 export default function RootLayout({
   children,
 }: {
   children: React.ReactNode
 }) {
   return (
-    <html lang="en">
-      {/*
-        <head /> will contain the components returned by the nearest parent
-        head.tsx. Find out more at https://beta.nextjs.org/docs/api-reference/file-conventions/head
-      */}
+    <html>
       <head />
-      <body>
-        <div>
-          Layout
-        </div>
-      </body>
+      <body>{children}</body>
     </html>
   )
 }
