+++ src/app/page.tsx
@@ -8,7 +8,7 @@ const inter = Inter({ subsets: ['latin'] })
 export default function Home() {
   return (
     <main className={styles.main}>
-      <div>Home</div>
+      <div>Hello</div>
     </main>
   )
 }
